files renamed to conventions of index.js and style.css. Comments added to components items
