+++ client/src/components/Footer/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import { useLocation, useHistory } from 'react-router-dom';
 
+// footer component for the app
+
 const Footer = () => {
   const location = useLocation();
   const history = useHistory();
