+++ client/src/components/Footer/style.css
@@ -1,0 +1,1 @@
+/* Will need to delete if unused for footer specific styling */
