+++ client/src/components/Header/style.css
@@ -1,0 +1,1 @@
+/* Will need to delete if unused for header specific styling */
