+++ client/src/components/MessageForm/style.css
@@ -1,0 +1,1 @@
+/* Will need to delete if unused for MessageForm specific styling */
