+++ client/src/components/ResponseForm/style.css
@@ -1,0 +1,1 @@
+/* Will need to delete if unused for ResponseForm specific styling */
