+++ client/src/components/ResponseList/style.css
@@ -1,0 +1,1 @@
+/* Will need to delete if unused for ResponseList specific styling */
