+++ client/src/components/SingleCategory/ProductOptionsCard/index.js
@@ -2,6 +2,7 @@ import React from "react";
 // import { useQuery } from '@apollo/client';
 import "./style.css";
 
+// This is the productOptionsCard component of the app. This is what will help us render the specific option per print category upon the on click of the link prompting the user in the Products page/product card//
 function ProductOptionsCard(props) {
   return (
     <div className="card">
