+++ client/src/components/SingleCategory/Wrapper/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import "./style.css";
 
+//this is just a wrapper to make the cards look contained and uniform
 function Wrapper(props) {
   return <div className="wrapper">{props.children}</div>;
 }
