+++ client/src/components/SingleCategory/Wrapper/style.css
